iPod breakout split from v2.6, new single-input board
The iPod connector board was pulled from V2.6 to send to OSHPark

v2.6 is pending revisions down from the ATMega 16U4 to [8|16|32]U2. May
shrink more.

A new board is in the works that can fit on the back of the stereo
connector, providing only an auxiliary input. Uses an ATTiny.